Fix index int problem

+++ chxanalys/chx_correlationc.py
@@ -66,7 +66,7 @@ def _one_time_process(buf, G, past_intensity_norm, future_intensity_norm,
     #maxqind=G.shape[1]
     for i in range(i_min, min(img_per_level[level], num_bufs)):
         # compute the index into the autocorrelation matrix
-        t_index = level * num_bufs / 2 + i
+        t_index = int( level * num_bufs / 2 + i )
         delay_no = (buf_no - i) % num_bufs
 
         # get the images for correlating
